Web-based
This is what I have for the web based application so far. I didn't use any Java, just JS, HTML, and Bootstrap CSS

You can bypass and view the apps as a guest. I don't have any functionalities included for when they login either as a user or admin yet. 

Also, I unfortunately am the only one with access to the database but it shouldn't be a big deal.